requirements.txt
av==15.1.0
loguru==0.7.3
mypy==1.18.2
opencv-python==4.12.0.88
pillow==11.3.0
pydantic-settings==2.11.0
pygrabber==0.2
pytest==8.4.2
python-jose==3.5.0
ruff==0.13.2
websockets==15.0.1